refactor: move K8s package to internal
