+++ internal/k8s/k8stest/k8stest.go
@@ -5,7 +5,7 @@ import (
 	"io"
 	"time"
 
-	"github.com/airbytehq/abctl/internal/cmd/local/k8s"
+	"github.com/airbytehq/abctl/internal/k8s"
 	v1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	networkingv1 "k8s.io/api/networking/v1"
